chore: move knex-specific loader logic out of EntityDataManager
