+++ packages/entity/src/internal/EntityKnexDataManager.ts
@@ -1,0 +1,84 @@
+import {
+  EntityDatabaseAdapter,
+  FieldEqualityCondition,
+  QuerySelectionModifiers,
+  QuerySelectionModifiersWithOrderByRaw,
+} from '../EntityDatabaseAdapter';
+import { EntityQueryContext } from '../EntityQueryContext';
+import { timeAndLogLoadEventAsync } from '../metrics/EntityMetricsUtils';
+import { EntityMetricsLoadType, IEntityMetricsAdapter } from '../metrics/IEntityMetricsAdapter';
+
+/**
+ * A knex data manager is responsible for handling non-dataloader-based
+ * database operations.
+ *
+ * @internal
+ */
+export class EntityKnexDataManager<
+  TFields extends Record<string, any>,
+  TIDField extends keyof TFields,
+> {
+  constructor(
+    private readonly databaseAdapter: EntityDatabaseAdapter<TFields, TIDField>,
+    private readonly metricsAdapter: IEntityMetricsAdapter,
+    private readonly entityClassName: string,
+  ) {}
+
+  /**
+   * Loads many objects matching the conjunction of where clauses constructed from
+   * specified field equality operands.
+   *
+   * @param queryContext - query context in which to perform the load
+   * @param fieldEqualityOperands - list of field equality where clause operand specifications
+   * @param querySelectionModifiers - limit, offset, and orderBy for the query
+   * @returns array of objects matching the query
+   */
+  async loadManyByFieldEqualityConjunctionAsync<N extends keyof TFields>(
+    queryContext: EntityQueryContext,
+    fieldEqualityOperands: FieldEqualityCondition<TFields, N>[],
+    querySelectionModifiers: QuerySelectionModifiers<TFields>,
+  ): Promise<readonly Readonly<TFields>[]> {
+    return await timeAndLogLoadEventAsync(
+      this.metricsAdapter,
+      EntityMetricsLoadType.LOAD_MANY_EQUALITY_CONJUNCTION,
+      this.entityClassName,
+      queryContext,
+    )(
+      this.databaseAdapter.fetchManyByFieldEqualityConjunctionAsync(
+        queryContext,
+        fieldEqualityOperands,
+        querySelectionModifiers,
+      ),
+    );
+  }
+
+  /**
+   * Loads many objects matching the raw WHERE clause.
+   *
+   * @param queryContext - query context in which to perform the load
+   * @param rawWhereClause - parameterized SQL WHERE clause with positional binding placeholders or named binding placeholders
+   * @param bindings - array of positional bindings or object of named bindings
+   * @param querySelectionModifiers - limit, offset, orderBy, and orderByRaw for the query
+   * @returns array of objects matching the query
+   */
+  async loadManyByRawWhereClauseAsync(
+    queryContext: EntityQueryContext,
+    rawWhereClause: string,
+    bindings: any[] | object,
+    querySelectionModifiers: QuerySelectionModifiersWithOrderByRaw<TFields>,
+  ): Promise<readonly Readonly<TFields>[]> {
+    return await timeAndLogLoadEventAsync(
+      this.metricsAdapter,
+      EntityMetricsLoadType.LOAD_MANY_RAW,
+      this.entityClassName,
+      queryContext,
+    )(
+      this.databaseAdapter.fetchManyByRawWhereClauseAsync(
+        queryContext,
+        rawWhereClause,
+        bindings,
+        querySelectionModifiers,
+      ),
+    );
+  }
+}
